+++ packages/plugins/src/index.ts
@@ -176,6 +176,9 @@ import * as cytonic from './plugins/cytonic';
 import * as bouncebit from './plugins/bouncebit';
 import * as haven from './plugins/haven';
 import * as pumpfun from './plugins/pumpfun';
+import * as helium from './plugins/helium';
+import * as carrot from './plugins/carrot';
+
 import { solanaSimpleFetcher } from './plugins/tokens';
 
 export { getFetchersByAddressSystem } from './utils/misc/getFetchersByAddressSystem';
@@ -364,6 +367,8 @@ const modules = [
   bouncebit,
   haven,
   pumpfun,
+  helium,
+  carrot,
 ];
 
 // JOBS //