+++ packages/plugins/src/plugins/helium/constants.ts
@@ -1,0 +1,10 @@
+import { PublicKey } from '@solana/web3.js';
+
+export const platformId = 'helium';
+
+export const programId = new PublicKey(
+  'hdaoVTCqhfHHo75XdAMxBKdUqvq1i5bF23sisBqVgGR'
+);
+export const stakeRegistryId = new PublicKey(
+  'hvsrNC3NKbcryqDs2DocYHZ9yPKEVzdSjQG6RVtK1s8'
+);