+++ packages/plugins/src/plugins/helium/depositFetcher.ts
@@ -1,0 +1,14 @@
+import { NetworkId } from '@sonarwatch/portfolio-core';
+import { Cache } from '../../Cache';
+import { platformId } from './constants';
+import { Fetcher, FetcherExecutor } from '../../Fetcher';
+
+const executor: FetcherExecutor = async (owner: string, cache: Cache) => [];
+
+const fetcher: Fetcher = {
+  id: `${platformId}-deposit`,
+  networkId: NetworkId.solana,
+  executor,
+};
+
+export default fetcher;