+++ packages/plugins/src/plugins/helium/index.ts
@@ -1,0 +1,6 @@
+import { Fetcher } from '../../Fetcher';
+import { Job } from '../../Job';
+import depositFetcher from './depositFetcher';
+
+export const jobs: Job[] = [];
+export const fetchers: Fetcher[] = [depositFetcher];